nsdl_cas/requirements.txt
pandas==2.2.3
plotly==6.0.1
PyMuPDF==1.25.5
streamlit==1.44.1
google-api-python-client==2.169.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
yfinance==0.2.55
